Clean-up of the auction contracts

+++ auction/src/auction/AuctionEnd.java
@@ -1,13 +1,15 @@
 package auction;
 
+import java.math.BigInteger;
+
 import takamaka.lang.Contract;
 import takamaka.lang.Event;
 
 public class AuctionEnd extends Event {
 	public final Contract highestBidder;
-	public final int highestBid;
+	public final BigInteger highestBid;
 
-	AuctionEnd(Contract highestBidder, int highestBid) {
+	AuctionEnd(Contract highestBidder, BigInteger highestBid) {
 		this.highestBidder = highestBidder;
 		this.highestBid = highestBid;
 	}
